+++ megatron/__init__.py
@@ -4,3 +4,4 @@ from . import io
 from . import nodes
 from . import layertools
 from .pipeline import Pipeline, load_pipeline
+from .layers import metrics